# openapi.yaml
openapi: 3.0.0
info:
  title: User and Mentor Management API
  version: 1.0.0
  description: API for managing users, mentors, offers, and authentication
servers:
  - url: 'https://prod-team-35-lg7sic6v.REDACTED/api'
paths:
...
  /admin/secreturl/mentor-requests/{request_id}:
    post:
      tags: [Admin]
      summary: Approve or reject a mentor request
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Approve'
      responses:
        '200':
          description: Request processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
...
  /admin/secreturl/mentor-requests:
    get:
      tags: [Admin]
      summary: Retrieve mentor requests
      parameters:
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            enum: [date, request_id]
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
        - name: date_from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: date_to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: group_by_status
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: A list of mentor requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorRequest'
...
components:
  schemas:
...
    MentorRequest:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        about:
          type: string
        direction:
          type: string
        date:
          type: string
          format: date-time
        status:
          type: boolean
...
    Approve:
      type: object
      properties:
        status:
          type: boolean
      required:
        - status
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT